+++ CHANGELOG.md
@@ -10,8 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.2] - 2025-08-31
 
 ### Added
+- exported utility function `make_reflection_agent` to provide boilerplate reflection agent (and optionally special instructions for how to improve the prompts)
+- new example in `examples/customer_support`
 
-- export utility `make_reflection_agent` function to provide boilerplate reflection agent
+### Changed
+- `Optimizer` calls more methods asynchronously to avoid nested event loops
+- Removed unused data types
 
 ## [0.0.1] - 2025-08-17